requirements.txt
aiohttp
nats-py
pyTelegramBotAPI
python-dotenv
pydantic
PyYaml
